requirements.txt
telethon~=1.36.0
aiogram~=3.7.0
psycopg2~=2.9.9
python-decouple~=3.8